\name{cinfoB.AB}
\alias{cinfoB.AB}
\docType{data}
\title{Vector indexing the matrix cmatB.AB}
\description{
This vector is used to describe the structure of the rows of the contrast matrix cmatB.AB
Its first element (1) indicates that the first esperimental question (main effect B) is described by a
single contrast in vectorial form. The second element (2)
indicates that the second experimental question (interaction between A and B) is given by
a contrast in matrix form.
}
\usage{Such a vector has to be created for every contrast martix.}
\format{ 
}
\details{
}
\source{
}
\references{
Bretz, F and Landgrebe J and Brunner E (2003):"Design and analysis of
        two colour factorial microarray experiments", submitted.
}
\examples{
data(cinfoB.AB)
}
\keyword{datasets}
