\name{data.3x2}
\alias{data.3x2}
\docType{data}
\title{3x2 microarray data}
\description{
These are data from a microarray
experiment in which the expression profiles of three cell lines were analysed
with and without drug treatment using cDNA microarrays spotted with 300012 human
cDNAs. The data matrix consists of 30012 rows and 18 columns. Each row represents one spot,
each column corresponds to one microarray. }
\usage{data(data.3x2)}
\format{
  The format is: matrix of size 30012 x 18.
}
\details{
}
\source{
cDNA microarray lab of the University of Goettingen, Germany.
\url{http://www.microarrays.med.uni-goettingen.de}
}
\references{
}
\examples{
data(data.3x2)
## maybe str(data.3x2) ; plot(data.3x2) ...
}
\keyword{datasets}
