\name{designs.basic}
\alias{designs.basic}
\docType{data}
\title{Basic designs for two-colour factorial 3 x 2 microarray data}
\description{ A list of matrices describing basic designs for two-colour factorial microarray data of
size 3 x 2. Matrix rows represent microarrays, matrix columns represent parameters. 
}
\usage{data(designs.basic)}
\format{
List of matrices of size 6 x 9.
}
\details{
The designs are abbreviated as in the paper (cf. source and references): 
BS - swap over B, AL - A loop,  XL - crossed loop,  CL - circle loop,  RS - star swap,  TL
- triangular loop, CR - common reference.
}
\source{
cDNA microarray lab of the University of Goettingen, Germany.
\url{http://www.microarrays.med.uni-goettingen.de}
}

\keyword{datasets}
