% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temp_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{gráfico de la temperatura mensual}
\usage{
grafico_temperatura_mensual(estacion, colores = NULL, titulo = "Temperatura")
}
\arguments{
\item{colores}{vector de colores para las series (opcional)}

\item{titulo}{título del gráfico. Por defecto "Temperatura"}

\item{estaciones}{dataframe con columnas "fecha", "temperatura_abrigo_150cm" y "id"}
}
\value{
un objeto ggplot
}
\description{
grafica la temperatura mensual por estación
}
\examples{
grafico_temperatura_mensual(estaciones, colores = c("tomato", "dodgerblue"), titulo = "Temperatura mensual promedio")

}
