\name{id.3x2}
\alias{id.3x2}
\docType{data}
\title{A vector of length 30012 containing numeric identifiers of the genes from the
microarray dataset data.3x2.}
\description{
Cf. data.3x2
}
\usage{data(id.3x2)}
\format{
  The format is:
 num [1:30012] 12 24 108 120 204 216 300 312 396 408 ...
}
\details{
}
\source{
}
\references{
}
\examples{
data(id.3x2)
}
\keyword{datasets}
