% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumen_temp.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{tabla resumen de temperatura}
\usage{
tabla_resumen_temperatura(estacion)
}
\arguments{
\item{estaciones}{dataframe con las columnas "id" y "temperatura_abrigo_150cm"}
}
\value{
un tibble con promedio y desvío estándar por estación
}
\description{
devuelve una tabla de resumen de temperatura_abrigo_150cm para una o más estaciones.
}
\examples{
df <- leer_estacion("NH0437", "datos/NH0437.csv")
  tabla_resumen_temperatura(df)

}
